Refactor: Rename fetchAndExtractContent to fetcher and remove comments
Based on your feedback:
- I renamed `fetchAndExtractContent` to `fetcher` in `useTools.ts` for brevity. This includes updating the tool definition, argument interface, and references in `executeToolCall`.
- I removed explanatory comments from `src/sidePanel/utils/scrapers.ts` and `src/sidePanel/hooks/useTools.ts`.
