+++ src/sidePanel/hooks/useTools.ts
@@ -2,6 +2,7 @@ import { useCallback } from 'react';
 import { useConfig } from '../ConfigContext';
 import { toast } from 'react-hot-toast';
 import { saveNoteInSystem } from 'src/background/noteStorage';
+import { scrapeUrlContent } from '../utils/scrapers';
 
 interface SaveNoteArgs {
   content: string;
@@ -10,7 +11,11 @@ interface SaveNoteArgs {
 }
 
 interface UpdateMemoryArgs {
-  summary: string; // The summary provided by the LLM
+  summary: string;
+}
+
+interface FetcherArgs {
+  url: string;
 }
 
 export interface LLMToolCall {
@@ -88,16 +93,9 @@ export const extractAndParseJsonArguments = (argsString: string): any => {
   throw new Error('Failed to parse arguments string as JSON after multiple attempts.');
 };
 
-/**
- * Custom hook for providing tools to the LLM.
- */
 export const useTools = () => {
   const { config, updateConfig } = useConfig();
-  /**
-   * Saves a note with the given content, title, and tags.
-   * Integrates with the existing note system for persistent storage.
-   * @param args An object containing the note's content, an optional title, and optional tags.
-   */
+
   const saveNote = useCallback(async (args: SaveNoteArgs): Promise<{ success: boolean; message: string }> => {
     const { content, title } = args;
     const llmTagsInput = args.tags;
@@ -130,11 +128,6 @@ export const useTools = () => {
     }
   }, []);
 
-  /**
-   * Appends a summary to the persistent popover note (memory).
-   * The summary is typically generated by the LLM based on user input.
-   * @param args An object containing the summary to append.
-   */
   const updateMemory = useCallback((args: UpdateMemoryArgs): { success: boolean; message: string } => {
     const { summary } = args;
 
@@ -172,7 +165,7 @@ export const useTools = () => {
         description: 'Saves a new note to the user\'s persistent note system. Use this when the user wants to record information, decisions, or create a new structured note.',
         parameters: {
           type: 'object',
-          structure: 'markdown', // Optional, can be used to indicate the format of the content
+          structure: 'markdown',
           properties: {
             content: {
               type: 'string',
@@ -185,12 +178,29 @@ export const useTools = () => {
             tags: {
               type: 'array',
               description: 'An optional list of tags (strings) to categorize the note.',
-              items: { // Note: OpenAI schema might represent array items differently, check specific LLM docs
+              items: {
                 type: 'string'
               }
             },
           },
           required: ['content'],
+        },
+      },
+    },
+    {
+      type: 'function',
+      function: {
+        name: 'fetcher',
+        description: 'Fetches the main textual content of a given URL. Use this to get the content of a webpage.',
+        parameters: {
+          type: 'object',
+          properties: {
+            url: {
+              type: 'string',
+              description: 'The URL of the webpage to fetch and extract content from.',
+            },
+          },
+          required: ['url'],
         },
       },
     },
@@ -213,11 +223,6 @@ export const useTools = () => {
     },
   ];
 
-  /**
-   * Executes a tool call requested by the LLM.
-   * @param toolCall The tool call object from the LLM.
-   * @returns A promise that resolves to the content to be sent back to the LLM.
-   */
   const executeToolCall = async (
     llmToolCall: LLMToolCall | { name: string; arguments: string; id?: string }
   ): Promise<{ toolCallId?: string; name: string; result: string }> => {
@@ -243,11 +248,17 @@ export const useTools = () => {
       } else if (toolName === 'updateMemory') {
         const { success, message } = updateMemory(args as UpdateMemoryArgs);
         return { toolCallId, name: toolName, result: message };
+      } else if (toolName === 'fetcher') {
+        const result = await scrapeUrlContent((args as FetcherArgs).url);
+        return { toolCallId, name: toolName, result };
       } else {
         return { toolCallId, name: toolName, result: `Error: Unknown tool '${toolName}'` };
       }
     } catch (error: any) {
       console.error(`Error executing tool ${toolName}:`, error);
+      if (toolName === 'fetcher' && typeof error === 'string' && error.startsWith('[Error scraping URL:')) {
+        return { toolCallId, name: toolName, result: error };
+      }
       return { toolCallId, name: toolName, result: `Error parsing arguments or executing tool ${toolName}: ${error.message}` };
     }
   };
